attach the notification icon to the exercises sidebar option
