+++ src/components/sidebar/NewSidebar.js
@@ -1,27 +1,17 @@
 import { useState, useEffect, useContext } from "react";
 import { useLocation } from "react-router-dom/cjs/react-router-dom";
-import { APIContext } from "../../contexts/APIContext";
 import { UserContext } from "../../contexts/UserContext";
-import { ExerciseCountContext } from "../../exercises/ExerciseCountContext";
-import { userHasNotExercisedToday } from "../../exercises/utils/daysSinceLastExercise";
 import SidebarOptions_Student from "./SidebarOptions_Student";
 import SidebarOptions_Teacher from "./SidebarOptions_Teacher";
-import NotificationIcon from "../NotificationIcon";
 import * as s from "./NewSidebar.sc";
 import NavOption from "./NavOption";
 import SettingsRoundedIcon from "@mui/icons-material/SettingsRounded";
 import FeedbackButton from "../FeedbackButton";
 import DoubleArrowRight from "@mui/icons-material/KeyboardDoubleArrowRightOutlined";
 import DoubleArrowLeft from "@mui/icons-material/KeyboardDoubleArrowLeftOutlined";
-import { MAX_EXERCISE_TO_DO_NOTIFICATION } from "../../exercises/ExerciseConstants";
 
 export default function NewSidebar({ isCollapsed, setIsCollapsed }) {
   const user = useContext(UserContext);
-  const api = useContext(APIContext);
-  const exerciseNotification = useContext(ExerciseCountContext);
-  const [hasExerciseNotification, setHasExerciseNotification] = useState(false);
-  const [totalExercisesInPipeline, setTotalExercisesInPipeline] = useState();
-
   const [isOnStudentSide, setIsOnStudentSide] = useState(true);
   const [isTeacher] = useState(user.is_teacher);
 
@@ -31,24 +21,6 @@ export default function NewSidebar({ isCollapsed, setIsCollapsed }) {
   useEffect(() => {
     setIsOnStudentSide(!path.includes("teacher"));
   }, [path]);
-
-  useEffect(() => {
-    if (userHasNotExercisedToday() && path !== "/exercises")
-      api.getUserBookmarksToStudy(1, (scheduledBookmaks) => {
-        exerciseNotification.setHasExercises(scheduledBookmaks.length > 0);
-        exerciseNotification.updateReactState();
-      });
-    else {
-      exerciseNotification.setHasExercises(false);
-      exerciseNotification.updateReactState();
-    }
-  }, [path]);
-
-  useEffect(() => {
-    exerciseNotification.setHasExercisesHook = setHasExerciseNotification;
-    exerciseNotification.setExerciseCounterHook = setTotalExercisesInPipeline;
-    exerciseNotification.updateReactState();
-  }, []);
 
   return (
     <s.SideBar
@@ -93,19 +65,6 @@ export default function NewSidebar({ isCollapsed, setIsCollapsed }) {
           isCollapsed={isCollapsed}
           text={"Settings"}
           currentPath={path}
-          notification={
-            !hasExerciseNotification && (
-              <NotificationIcon
-                text={
-                  totalExercisesInPipeline
-                    ? totalExercisesInPipeline > MAX_EXERCISE_TO_DO_NOTIFICATION
-                      ? MAX_EXERCISE_TO_DO_NOTIFICATION + "+"
-                      : totalExercisesInPipeline
-                    : ""
-                }
-              />
-            )
-          }
         />
 
         <FeedbackButton isCollapsed={isCollapsed} />
