+++ src/components/sidebar/SidebarOptions_Student.js
@@ -1,3 +1,10 @@
+import { useState, useEffect, useContext } from "react";
+import { useLocation } from "react-router-dom/cjs/react-router-dom";
+import { APIContext } from "../../contexts/APIContext";
+import { ExerciseCountContext } from "../../exercises/ExerciseCountContext";
+import { userHasNotExercisedToday } from "../../exercises/utils/daysSinceLastExercise";
+import { MAX_EXERCISE_TO_DO_NOTIFICATION } from "../../exercises/ExerciseConstants";
+import NotificationIcon from "../NotificationIcon";
 import NavOption from "./NavOption";
 import HomeRoundedIcon from "@mui/icons-material/HomeRounded";
 import FitnessCenterRoundedIcon from "@mui/icons-material/FitnessCenterRounded";
@@ -11,6 +18,29 @@ export default function SidebarOptions_Student({
   currentPath,
   isTeacher,
 }) {
+  const api = useContext(APIContext);
+  const exerciseNotification = useContext(ExerciseCountContext);
+  const [hasExerciseNotification, setHasExerciseNotification] = useState(false);
+  const [totalExercisesInPipeline, setTotalExercisesInPipeline] = useState();
+  const path = useLocation().pathname;
+
+  useEffect(() => {
+    if (userHasNotExercisedToday() && path !== "/exercises")
+      api.getUserBookmarksToStudy(1, (scheduledBookmaks) => {
+        exerciseNotification.setHasExercises(scheduledBookmaks.length > 0);
+        exerciseNotification.updateReactState();
+      });
+    else {
+      exerciseNotification.setHasExercises(false);
+      exerciseNotification.updateReactState();
+    }
+  }, [path]);
+
+  useEffect(() => {
+    exerciseNotification.setHasExercisesHook = setHasExerciseNotification;
+    exerciseNotification.setExerciseCounterHook = setTotalExercisesInPipeline;
+    exerciseNotification.updateReactState();
+  }, []);
   return (
     <>
       <NavOption
@@ -27,6 +57,19 @@ export default function SidebarOptions_Student({
         isCollapsed={isCollapsed}
         text={"Exercises"}
         currentPath={currentPath}
+        notification={
+          hasExerciseNotification && (
+            <NotificationIcon
+              text={
+                totalExercisesInPipeline
+                  ? totalExercisesInPipeline > MAX_EXERCISE_TO_DO_NOTIFICATION
+                    ? MAX_EXERCISE_TO_DO_NOTIFICATION + "+"
+                    : totalExercisesInPipeline
+                  : ""
+              }
+            />
+          )
+        }
       />
 
       <NavOption
